requirements.txt
openai
langchain
sentence-transformers
tiktoken
pdfminer.six
beautifulsoup4
scrapy